// src/shared/api/generated/models/templateCreateDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * SUM RM API
 * OpenAPI spec version: 1.0
 */
import { ColumnState } from 'ag-grid-community';
import type { TemplateCreateDtoFilterModel } from './templateCreateDtoFilterModel';

export interface TemplateCreateDto {
  template_name: string;
  public: boolean;
  filterModel: TemplateCreateDtoFilterModel;
  sortState?: string[];
  columnState?: string[] | ColumnState[];
  selectedIds?: string[];
}