// src/shared/api/generated/models/templateUpdateDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * SUM RM API
 * OpenAPI spec version: 1.0
 */
import { ColumnState } from 'ag-grid-community';
import type { TemplateUpdateDtoFilterModel } from './templateUpdateDtoFilterModel';

export interface TemplateUpdateDto {
  template_id: number;
  template_name?: string;
  public?: boolean;
  filterModel?: TemplateUpdateDtoFilterModel;
  sortState?: string[];
  columnState?: string[] | ColumnState[];
  selectedIds?: string[];
}